use fixed type for 64

+++ src/utils.c
@@ -4,12 +4,12 @@
 
 #define US_PER_SECOND 1000000
 
-unsigned long long int time_to_us(struct timeval tv)
+uint64_t time_to_us(struct timeval tv)
 {
     return tv.tv_sec * US_PER_SECOND + tv.tv_usec;
 }
 
-unsigned long long int get_time_us()
+uint64_t get_time_us()
 {
     struct timeval tv;
     gettimeofday(&tv, NULL);
